JupyterNotebookSupervisedLearningProject-Classification
This was a project in the udemy Data Science and Machine Learning 2022 boot-camp. We used Jupyter Notebooks with the tools Pandas, Numpy, Scikit-Learn, MatPlotLib, and seaborn. We experimented with different estimators to see which worked best with our data. We found LogisticRegression came closest to our accuracy goal, though even with tuning was at 88% instead of 95%.